refactor

+++ server/core/store_option.go
@@ -164,7 +164,7 @@ func SetAvailableFunc(f func() bool) StoreCreateOption {
 }
 
 // SetMaxScore is used to set maxScore
-func SetMaxScore(maxScore uint64) StoreCreateOption {
+func SetMaxScore(maxScore float64) StoreCreateOption {
 	return func(store *StoreInfo) {
 		store.maxScore = maxScore
 	}
